pyproject.toml
[project]
name = "temporal-community-detection"
version = "0.1.0"
description = "Detecting evolving communities in continuous-time dynamic graphs"
readme = "README.md"
license = {text = "GPL"}
authors = [
    {name = "Morteza Ziabakhsh", email = "morteza24mail@protonmail.com"},
]
keywords = [
    "temporal-graphs",
    "dynamic-graphs",
    "community-detection",
    "graph-neural-networks",
    "temporal-gnn",
    "enron",
    "tgb",
]

dependencies = [
    "torch",
    "torch_geometric",
    "py-tgb",
    "numpy",
    "scikit-learn",
    "pandas",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["tcd*"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
